Refactor code to properly check for wifi connection
