+++ src/wifi-setup/wait.js
@@ -1,5 +1,5 @@
 module.exports = function wait(milliseconds) {
-  return new Promise(function(resolve, reject) {
+  return new Promise(function(resolve) {
     setTimeout(resolve, milliseconds);
   });
 };
