+++ src/wifi-setup/wifi.js
@@ -13,6 +13,13 @@ exports.removeNetwork = removeNetwork;
 exports.getKnownNetworks = getKnownNetworks;
 exports.broadcastBeacon = broadcastBeacon;
 exports.checkConnection = checkConnection;
+exports.waitForWiFi = waitForWiFi;
+
+// The Edison device can't scan for wifi networks while in AP mode, so
+// we've got to scan before we enter AP mode and save the results. They're a
+// global variable because this code needs to ship
+exports.preliminaryScanResults = [];
+
 
 /**
  * Determine whether we have a wifi connection with the `wpa_cli
@@ -239,5 +246,100 @@ function checkConnection() {
   if (fs.existsSync('wifiskip')) {
     return Promise.resolve(true);
   }
-  return Promise.resolve(false);
+  return getStatus().then(() => {
+    // Wait until we have a working wifi connection. Retry every 3 seconds up
+    // to 10 times. If we are connected, then start the Gateway client.
+    // If we never get a wifi connection, go into AP mode.
+    // Before we start, though, let the user know that something is happening
+    waitForWiFi(20, 3000).then(() => {
+      return true;
+    }).catch((err) => {
+      console.log('No wifi connection found. Starting the AP...', err);
+      // Scan for wifi networks now because we can't always scan once
+      // the AP is being broadcast, retrying up to 10 times
+      scan(10).then((ssids) => {
+        exports.preliminaryScanResults = ssids; // Remember ssids
+        console.log('No wifi found; entering AP mode');
+        startAP(platform.ap_ip);
+      });
+
+      return false;
+    });
+  }).catch((err) => {
+    console.error('Error checking wifi adapter presence', err);
+    return true;
+  });
+}
+
+// Return a promise, then check every interval ms for a wifi connection.
+// Resolve the promise when we're connected. Or, if we aren't connected
+// after maxAttempts attempts, then reject the promise
+function waitForWiFi(maxAttempts, interval) {
+  return new Promise(function(resolve, reject) {
+    let attempts = 0;
+
+    // first of all we query wpa_supplicant if there's a wifi AP configured
+    run(platform.listNetworks)
+      .then((out) => {
+        console.log('List Networks command executed:', out);
+        if (out.includes('\n0\t')) {
+          // there's at least one wifi AP configured. Let's wait to see if it
+          // will connect
+          check();
+        } else {
+          // No wifi AP configured. Let's skip the wait and start the setup
+          // immediately
+          reject();
+        }
+      })
+      .catch((err) => console.error('Error listing Networks:', err));
+
+
+    function check() {
+      attempts++;
+      console.log('check', attempts);
+      getStatus()
+        .then((status) => {
+          console.log(status);
+          if (status === 'COMPLETED') {
+            console.log('WiFi connection found. resolving');
+            checkForAddress();
+            console.log('resolved');
+          } else {
+            console.log('No wifi connection on attempt', attempts);
+            retryOrGiveUp();
+          }
+        })
+        .catch((err) => {
+          console.error('Error checking wifi on attempt', attempts, ':', err);
+          retryOrGiveUp();
+        });
+    }
+
+    function checkForAddress() {
+      const ifaces = os.networkInterfaces();
+
+      if (ifaces.hasOwnProperty('wlan0')) {
+        for (const addr of ifaces.wlan0) {
+          if (addr.family !== 'IPv4' || addr.internal) {
+            continue;
+          }
+
+          resolve();
+          return;
+        }
+      }
+
+      retryOrGiveUp();
+    }
+
+    function retryOrGiveUp() {
+      if (attempts >= maxAttempts) {
+        console.error('Giving up. No wifi available.');
+        reject();
+      } else {
+        setTimeout(check, interval);
+      }
+    }
+  });
 }
